yolo

+++ NEAT/NEATModels/neat_faststandard.py
@@ -204,7 +204,7 @@ class NEATDynamicSegFree(object):
         
         self.maskimagename = maskimagename
         self.maskimage = imread(maskimagename)
-        self.heatmap = np.zeros(self.image.shape, dtype = 'float16')
+        self.heatmap = np.zeros(self.image.shape, dtype = 'float32')
         self.savedir = savedir
         self.n_tiles = n_tiles
         self.thresh = thresh
